fix: cast numeric fields to float in insights, fix all API routes

+++ frontend/src/app/api/insights/route.ts
@@ -1,5 +1,6 @@
 import { NextRequest, NextResponse } from 'next/server'
 import { queryOne } from '../db'
+
 export async function GET(req: NextRequest) {
   const category = req.nextUrl.searchParams.get('category')
   const state = req.nextUrl.searchParams.get('state')
@@ -9,10 +10,11 @@ export async function GET(req: NextRequest) {
   if (state) { params.push(state.toUpperCase()); conditions.push(`p.state=$${params.length}`) }
   const where = conditions.length ? 'AND ' + conditions.join(' AND ') : ''
   const summary = await queryOne(`
-    SELECT COUNT(DISTINCT a.plant_id) AS plants_with_attribution,
-      COUNT(DISTINCT a.plant_id) FILTER (WHERE a.pct_change_total < 0) AS plants_improving,
-      COUNT(DISTINCT a.plant_id) FILTER (WHERE a.pct_change_total > 0) AS plants_worsening,
-      ROUND(AVG(a.pct_change_total)::numeric,1) AS avg_gwp_change_pct
+    SELECT 
+      COUNT(DISTINCT a.plant_id)::int AS plants_with_attribution,
+      COUNT(DISTINCT a.plant_id) FILTER (WHERE a.pct_change_total < 0)::int AS plants_improving,
+      COUNT(DISTINCT a.plant_id) FILTER (WHERE a.pct_change_total > 0)::int AS plants_worsening,
+      ROUND(AVG(a.pct_change_total)::numeric,1)::float AS avg_gwp_change_pct
     FROM gwp_attribution a JOIN plants p ON a.plant_id=p.id
     WHERE a.pct_change_total IS NOT NULL ${where}
   `, params)
